fix: serve progress display helpers with renderer
